+++ src/components/Cart/Cart.js
@@ -1,0 +1,12 @@
+import React from "react";
+import CartPopupModal from "./CartPopupModal";
+import FloatingCartButton from "./FloatingCartButton";
+
+export default function Cart({ cart, isCartOpen, toggleCart }) {
+  return (
+    <>
+      <CartPopupModal cart={cart} isCartOpen={isCartOpen} />
+      <FloatingCartButton toggleCart={toggleCart} />
+    </>
+  );
+}
