+++ src/components/Cart/FloatingCartButton.js
@@ -1,0 +1,18 @@
+import React from "react";
+import Button from "components/Button/Button";
+import "./Cart.css";
+
+export default function FloatingCartButton({ toggleCart }) {
+  return (
+    <>
+      <Button id="floatingCartButton" btnOnClick={toggleCart}>
+        <img
+          id="toggleIcon"
+          src="assets/cart.svg"
+          style={{ width: "30px" }}
+          alt="hamburger Icon"
+        />
+      </Button>
+    </>
+  );
+}
