+++ src/pages/order-now/PizzaList/OrderCard/ProductPrice.js
@@ -1,0 +1,6 @@
+import React from "react";
+import "../../ordernow.css";
+
+export default function ProductPrice({ quantity, price }) {
+  return <p className="price fs-4 mb-0">{`$${quantity * price}`}</p>;
+}
